fix/examples: don't prematurely drop Service
Updates example to keep Service and Listener objects alive while example is running. Otherwise connections will be dropped and example won't work as expected.
